requirements.txt
# MCP framework
fastmcp

# PostgreSQL async driver (psycopg v3, binary build — no system libpq needed)
psycopg[binary]

# Environment variable management
python-dotenv

# HTTP server
fastapi
uvicorn[standard]

# JWT authentication
python-jose[cryptography]

# Password hashing
bcrypt
